smart contracts 111 - testing works, prep largeAmountsTest
