+++ smartContract/artifacts/hardhat/console.sol/console.dbg.json
@@ -1,4 +1,4 @@
 {
   "_format": "hh-sol-dbg-1",
-  "buildInfo": "..\\..\\build-info\\ea9bb1a805b2a5d4c47b8b7fcb7c85f5.json"
+  "buildInfo": "..\\..\\build-info\\0a66c394dfd06938465d28288047a8be.json"
 }
